+++ handright/__init__.py
@@ -1,12 +1,9 @@
 # coding: utf-8
 """A lightweight Python library for simulating Chinese handwriting
 
-Vision: Reveal the nature of Chinese handwriting and use it to implement
-beautiful, simple and easy-to-use interfaces.
+Created by Chenghui Li (Gsllchb).
 
-Algorithm: Randomly perturb each character as a whole in horizontal position,
-vertical position and font size. Then, Randomly perturb each stroke of a
-character in horizontal position, vertical position and rotation angle.
+Licensed under BSD 3-Clause.
 """
 from handright._core import (
     handwrite, Template, Error, LayoutError, BackgroundTooLargeError