+++ docs/release_notes.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## unreleased
+* 修改Handright库docstring的描述。
+* 更新作者信息。
+
 ## v5.2.0 (2019-10-20)
 * 优化function annotations。
 * 类`Template`添加`release_font_resource`和`__repr__`方法，`__slots__`属性。